bumped files changed through moving/deleting
