+++ src/main/kotlin/org/developerden/codosseum/config/ChallengesConfiguration.kt
@@ -1,28 +1,15 @@
 package org.developerden.codosseum.config
 
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.decodeFromStream
-import org.developerden.codosseum.files.git.Repository
+import org.developerden.codosseum.challenge.update.git.Repository
+import java.nio.file.Path
 import java.nio.file.Paths
-import kotlin.io.path.createFile
-import kotlin.io.path.inputStream
-import kotlin.io.path.notExists
+import kotlin.time.Duration
 import kotlin.time.Duration.Companion.minutes
-import kotlin.time.DurationUnit
 
-@Serializable
-data class ChallengesConfiguration(
-	val locals: Collection<String> = emptyList(),
-	val remoteUpdatePeriod: Long = 10.minutes.toLong(DurationUnit.MINUTES),
-	val repositories: Collection<Repository> = emptyList()
-) {
+object ChallengesConfiguration {
 
-	companion object {
-		@OptIn(ExperimentalSerializationApi::class)
-		fun loadConfig(): ChallengesConfiguration {
-			return Json.decodeFromStream(Paths.get("./challenges.json").apply { if (notExists()) createFile() }.inputStream())
-		}
-	}
+	val localPaths: Collection<Path> = System.getenv("LOCAL_PATHS").split(";").map(Paths::get)
+	val remoteUpdatePeriodMinutes: Duration = System.getenv("REMOTE_UPDATE_PERIOD_MINUTES").toInt().minutes
+	val repositories: Collection<Repository> = System.getenv("GIT_REPOSITORIES").split(";").map(Repository::fromUrl)
+
 }
